requirements.txt
psycopg2 
djangorestframework 
django-cors-headers
django  
djangorestframework-simplejwt
djoser
